+++ src/main/java/duke/DukeException.java
@@ -2,8 +2,8 @@ package duke;
 
 public class DukeException extends Exception {
 
-    public DukeException(String message) {
-        super(message);
-    }
+	public DukeException(String message) {
+		super(message);
+	}
 
 }
